+++ app/src/main/java/reikai/presentation/migrate/flow/NovelMigrationFlowAdapter.kt
@@ -275,6 +275,7 @@ class NovelMigrationFlowAdapter(
                     novelRepository,
                     database,
                     libraryPreferences,
+                    coverCache,
                     novelDownloadManager = downloadManager,
                 )
             }